Updates for moving code examples to the Code Example Catalog.
