+++ javascript/example_code/mediaconvert/emc_createjob.js
@@ -23,18 +23,153 @@
 
 // ABOUT THIS NODE.JS SAMPLE: This sample is part of the SDK for JavaScript Developer Guide topic at
 // https://docs.aws.amazon.com/sdk-for-javascript/v2/developer-guide/emc-examples-jobs.html
-//
-// NOTE:
-// This is part 3 of 3 for this example.
-// The first part is emc_createjob_config.js.
-// The second part is emc_createjob_define.js.
-// The third part is emc_createjob.js. (this file)
-//
 
 // snippet-start:[mediaconvert.JavaScript.jobs.createJob]
-//
-// Part 3 of 3
-//
+
+// snippet-start:[mediaconvert.JavaScript.jobs.createJob_config]
+// Load the SDK for JavaScript
+var AWS = require('aws-sdk');
+// Set the region 
+AWS.config.update({region: 'us-west-2'});
+// Set the custom endpoint for your acccount
+AWS.config.mediaconvert({endpoint: 'ACCOUNT_ENDPOINT'});
+// snippet-end:[mediaconvert.JavaScript.jobs.createJob_config]
+
+// snippet-start:[mediaconvert.JavaScript.jobs.createJob_define]
+var params = {
+  "Queue": "JOB_QUEUE_ARN",
+  "UserMetadata": {
+    "Customer": "Amazon"
+  },
+  "Role": "IAM_ROLE_ARN",
+  "Settings": {
+    "OutputGroups": [
+      {
+        "Name": "File Group",
+        "OutputGroupSettings": {
+          "Type": "FILE_GROUP_SETTINGS",
+          "FileGroupSettings": {
+            "Destination": "s3://OUTPUT_BUCKET_NAME/"
+          }
+        },
+        "Outputs": [
+          {
+            "VideoDescription": {
+              "ScalingBehavior": "DEFAULT",
+              "TimecodeInsertion": "DISABLED",
+              "AntiAlias": "ENABLED",
+              "Sharpness": 50,
+              "CodecSettings": {
+                "Codec": "H_264",
+                "H264Settings": {
+                  "InterlaceMode": "PROGRESSIVE",
+                  "NumberReferenceFrames": 3,
+                  "Syntax": "DEFAULT",
+                  "Softness": 0,
+                  "GopClosedCadence": 1,
+                  "GopSize": 90,
+                  "Slices": 1,
+                  "GopBReference": "DISABLED",
+                  "SlowPal": "DISABLED",
+                  "SpatialAdaptiveQuantization": "ENABLED",
+                  "TemporalAdaptiveQuantization": "ENABLED",
+                  "FlickerAdaptiveQuantization": "DISABLED",
+                  "EntropyEncoding": "CABAC",
+                  "Bitrate": 5000000,
+                  "FramerateControl": "SPECIFIED",
+                  "RateControlMode": "CBR",
+                  "CodecProfile": "MAIN",
+                  "Telecine": "NONE",
+                  "MinIInterval": 0,
+                  "AdaptiveQuantization": "HIGH",
+                  "CodecLevel": "AUTO",
+                  "FieldEncoding": "PAFF",
+                  "SceneChangeDetect": "ENABLED",
+                  "QualityTuningLevel": "SINGLE_PASS",
+                  "FramerateConversionAlgorithm": "DUPLICATE_DROP",
+                  "UnregisteredSeiTimecode": "DISABLED",
+                  "GopSizeUnits": "FRAMES",
+                  "ParControl": "SPECIFIED",
+                  "NumberBFramesBetweenReferenceFrames": 2,
+                  "RepeatPps": "DISABLED",
+                  "FramerateNumerator": 30,
+                  "FramerateDenominator": 1,
+                  "ParNumerator": 1,
+                  "ParDenominator": 1
+                }
+              },
+              "AfdSignaling": "NONE",
+              "DropFrameTimecode": "ENABLED",
+              "RespondToAfd": "NONE",
+              "ColorMetadata": "INSERT"
+            },
+            "AudioDescriptions": [
+              {
+                "AudioTypeControl": "FOLLOW_INPUT",
+                "CodecSettings": {
+                  "Codec": "AAC",
+                  "AacSettings": {
+                    "AudioDescriptionBroadcasterMix": "NORMAL",
+                    "RateControlMode": "CBR",
+                    "CodecProfile": "LC",
+                    "CodingMode": "CODING_MODE_2_0",
+                    "RawFormat": "NONE",
+                    "SampleRate": 48000,
+                    "Specification": "MPEG4",
+                    "Bitrate": 64000
+                  }
+                },
+                "LanguageCodeControl": "FOLLOW_INPUT",
+                "AudioSourceName": "Audio Selector 1"
+              }
+            ],
+            "ContainerSettings": {
+              "Container": "MP4",
+              "Mp4Settings": {
+                "CslgAtom": "INCLUDE",
+                "FreeSpaceBox": "EXCLUDE",
+                "MoovPlacement": "PROGRESSIVE_DOWNLOAD"
+              }
+            },
+            "NameModifier": "_1"
+          }
+        ]
+      }
+    ],
+    "AdAvailOffset": 0,
+    "Inputs": [
+      {
+        "AudioSelectors": {
+          "Audio Selector 1": {
+            "Offset": 0,
+            "DefaultSelection": "NOT_DEFAULT",
+            "ProgramSelection": 1,
+            "SelectorType": "TRACK",
+            "Tracks": [
+              1
+            ]
+          }
+        },
+        "VideoSelector": {
+          "ColorSpace": "FOLLOW"
+        },
+        "FilterEnable": "AUTO",
+        "PsiControl": "USE_PSI",
+        "FilterStrength": 0,
+        "DeblockFilter": "DISABLED",
+        "DenoiseFilter": "DISABLED",
+        "TimecodeSource": "EMBEDDED",
+        "FileInput": "s3://INPUT_BUCKET_AND_FILE_NAME"
+      }
+    ],
+    "TimecodeConfig": {
+      "Source": "EMBEDDED"
+    }
+  }
+};
+// snippet-end:[mediaconvert.JavaScript.jobs.createJob_define]
+
+// snippet-start:[mediaconvert.JavaScript.jobs.createJob_create]
 // Create a promise on a MediaConvert object
 var endpointPromise = new AWS.MediaConvert({apiVersion: '2017-08-29'}).createJob(params).promise();
 
@@ -47,4 +182,6 @@ endpointPromise.then(
     console.log("Error", err);
   }
 );
+// snippet-end:[mediaconvert.JavaScript.jobs.createJob_create]
+
 // snippet-end:[mediaconvert.JavaScript.jobs.createJob]
