--- javascript/example_code/mediaconvert/emc_createjob_define.js
@@ -1,169 +1,0 @@
-/**
- * Copyright 2010-2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
- *
- * This file is licensed under the Apache License, Version 2.0 (the "License").
- * You may not use this file except in compliance with the License. A copy of
- * the License is located at
- *
- * http://aws.amazon.com/apache2.0/
- *
- * This file is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
- * CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License.
-*/
-
-//snippet-sourcedescription:[emc_createjob.js demonstrates how to create a transcoding job.]
-//snippet-service:[mediaconvert]
-//snippet-keyword:[JavaScript]
-//snippet-keyword:[Code Sample]
-//snippet-keyword:[AWS Elemental MediaConvert]
-//snippet-sourcetype:[full-example]
-//snippet-sourcedate:[2018-06-02]
-//snippet-sourceauthor:[AWS-JSDG]
-
-// ABOUT THIS NODE.JS SAMPLE: This sample is part of the SDK for JavaScript Developer Guide topic at
-// https://docs.aws.amazon.com/sdk-for-javascript/v2/developer-guide/emc-examples-jobs.html
-//
-// NOTE:
-// This is part 2 of 3 for this example.
-// The first part is emc_createjob_config.js.
-// The second part is emc_createjob_define.js. (this file)
-// The third part is emc_createjob.js.
-//
-
-// snippet-start:[mediaconvert.JavaScript.jobs.createJob_define]
-//
-// Part 2 of 3
-//
-var params = {
-  "Queue": "JOB_QUEUE_ARN",
-  "UserMetadata": {
-    "Customer": "Amazon"
-  },
-  "Role": "IAM_ROLE_ARN",
-  "Settings": {
-    "OutputGroups": [
-      {
-        "Name": "File Group",
-        "OutputGroupSettings": {
-          "Type": "FILE_GROUP_SETTINGS",
-          "FileGroupSettings": {
-            "Destination": "s3://OUTPUT_BUCKET_NAME/"
-          }
-        },
-        "Outputs": [
-          {
-            "VideoDescription": {
-              "ScalingBehavior": "DEFAULT",
-              "TimecodeInsertion": "DISABLED",
-              "AntiAlias": "ENABLED",
-              "Sharpness": 50,
-              "CodecSettings": {
-                "Codec": "H_264",
-                "H264Settings": {
-                  "InterlaceMode": "PROGRESSIVE",
-                  "NumberReferenceFrames": 3,
-                  "Syntax": "DEFAULT",
-                  "Softness": 0,
-                  "GopClosedCadence": 1,
-                  "GopSize": 90,
-                  "Slices": 1,
-                  "GopBReference": "DISABLED",
-                  "SlowPal": "DISABLED",
-                  "SpatialAdaptiveQuantization": "ENABLED",
-                  "TemporalAdaptiveQuantization": "ENABLED",
-                  "FlickerAdaptiveQuantization": "DISABLED",
-                  "EntropyEncoding": "CABAC",
-                  "Bitrate": 5000000,
-                  "FramerateControl": "SPECIFIED",
-                  "RateControlMode": "CBR",
-                  "CodecProfile": "MAIN",
-                  "Telecine": "NONE",
-                  "MinIInterval": 0,
-                  "AdaptiveQuantization": "HIGH",
-                  "CodecLevel": "AUTO",
-                  "FieldEncoding": "PAFF",
-                  "SceneChangeDetect": "ENABLED",
-                  "QualityTuningLevel": "SINGLE_PASS",
-                  "FramerateConversionAlgorithm": "DUPLICATE_DROP",
-                  "UnregisteredSeiTimecode": "DISABLED",
-                  "GopSizeUnits": "FRAMES",
-                  "ParControl": "SPECIFIED",
-                  "NumberBFramesBetweenReferenceFrames": 2,
-                  "RepeatPps": "DISABLED",
-                  "FramerateNumerator": 30,
-                  "FramerateDenominator": 1,
-                  "ParNumerator": 1,
-                  "ParDenominator": 1
-                }
-              },
-              "AfdSignaling": "NONE",
-              "DropFrameTimecode": "ENABLED",
-              "RespondToAfd": "NONE",
-              "ColorMetadata": "INSERT"
-            },
-            "AudioDescriptions": [
-              {
-                "AudioTypeControl": "FOLLOW_INPUT",
-                "CodecSettings": {
-                  "Codec": "AAC",
-                  "AacSettings": {
-                    "AudioDescriptionBroadcasterMix": "NORMAL",
-                    "RateControlMode": "CBR",
-                    "CodecProfile": "LC",
-                    "CodingMode": "CODING_MODE_2_0",
-                    "RawFormat": "NONE",
-                    "SampleRate": 48000,
-                    "Specification": "MPEG4",
-                    "Bitrate": 64000
-                  }
-                },
-                "LanguageCodeControl": "FOLLOW_INPUT",
-                "AudioSourceName": "Audio Selector 1"
-              }
-            ],
-            "ContainerSettings": {
-              "Container": "MP4",
-              "Mp4Settings": {
-                "CslgAtom": "INCLUDE",
-                "FreeSpaceBox": "EXCLUDE",
-                "MoovPlacement": "PROGRESSIVE_DOWNLOAD"
-              }
-            },
-            "NameModifier": "_1"
-          }
-        ]
-      }
-    ],
-    "AdAvailOffset": 0,
-    "Inputs": [
-      {
-        "AudioSelectors": {
-          "Audio Selector 1": {
-            "Offset": 0,
-            "DefaultSelection": "NOT_DEFAULT",
-            "ProgramSelection": 1,
-            "SelectorType": "TRACK",
-            "Tracks": [
-              1
-            ]
-          }
-        },
-        "VideoSelector": {
-          "ColorSpace": "FOLLOW"
-        },
-        "FilterEnable": "AUTO",
-        "PsiControl": "USE_PSI",
-        "FilterStrength": 0,
-        "DeblockFilter": "DISABLED",
-        "DenoiseFilter": "DISABLED",
-        "TimecodeSource": "EMBEDDED",
-        "FileInput": "s3://INPUT_BUCKET_AND_FILE_NAME"
-      }
-    ],
-    "TimecodeConfig": {
-      "Source": "EMBEDDED"
-    }
-  }
-};
-// snippet-end:[mediaconvert.JavaScript.jobs.createJob_define]
